requirements.txt
flask==3.1.3
flask-cors==6.0.2
gunicorn==25.1.0
requests==2.32.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
asyncpg==0.29.0
pgqueuer==0.1.0
